requirements.txt
streamlit
pandas
requests
beautifulsoup4
google-auth
google-cloud-aiplatform